Refactors the PredPatt module by restructuring the `__init__.py`, `corpus.py`, and `graph.py` files. Enhances documentation with detailed descriptions of classes and methods, improving clarity and usability. Introduces the `PredPattCorpus` and `PredPattGraphBuilder` classes for better management of semantic extractions and graph construction. Updates type hints for improved type safety and consistency across the module.
